rename pgsql to pg
